+++ fhir_auth/demos/mobileauthdemo/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'aidbox_request.dart';
+// import 'aidbox_request.dart';
 import 'gcs_request.dart';
 import 'hapi_request.dart';
 import 'meld_request.dart';
@@ -29,14 +29,14 @@ class MyApp extends StatelessWidget {
                     ),
                     child: Image.asset('assets/hapi.png'),
                     onPressed: () async => await hapiRequest()),
-                ElevatedButton(
-                    style: ButtonStyle(
-                      backgroundColor: MaterialStateProperty.all(Colors.white),
-                      side: MaterialStateProperty.all(
-                          const BorderSide(color: Colors.black)),
-                    ),
-                    child: Image.asset('assets/aidbox.png'),
-                    onPressed: () async => await aidboxRequest()),
+                // ElevatedButton(
+                //     style: ButtonStyle(
+                //       backgroundColor: MaterialStateProperty.all(Colors.white),
+                //       side: MaterialStateProperty.all(
+                //           const BorderSide(color: Colors.black)),
+                //     ),
+                //     child: Image.asset('assets/aidbox.png'),
+                //     onPressed: () async => await aidboxRequest()),
                 ElevatedButton(
                     style: ButtonStyle(
                       backgroundColor: MaterialStateProperty.all(Colors.white),
@@ -45,14 +45,14 @@ class MyApp extends StatelessWidget {
                     ),
                     child: Image.asset('assets/meld.png'),
                     onPressed: () async => await meldRequest()),
-                ElevatedButton(
-                    style: ButtonStyle(
-                      backgroundColor: MaterialStateProperty.all(Colors.white),
-                      side: MaterialStateProperty.all(
-                          const BorderSide(color: Colors.black)),
-                    ),
-                    child: Image.asset('assets/azure.png'),
-                    onPressed: () {}),
+                // ElevatedButton(
+                //     style: ButtonStyle(
+                //       backgroundColor: MaterialStateProperty.all(Colors.white),
+                //       side: MaterialStateProperty.all(
+                //           const BorderSide(color: Colors.black)),
+                //     ),
+                //     child: Image.asset('assets/azure.png'),
+                //     onPressed: () {}),
                 ElevatedButton(
                     style: ButtonStyle(
                       backgroundColor: MaterialStateProperty.all(Colors.white),