+++ fhir_path/test/webtest/macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/opt/homebrew/Caskroom/flutter/3.0.5/flutter"
-export "FLUTTER_APPLICATION_PATH=/Users/juun/workspace/fhir/fhir_path/test/webtest"
+export "FLUTTER_ROOT=/home/grey/snap/flutter/common/flutter"
+export "FLUTTER_APPLICATION_PATH=/home/grey/dev/fhir/fhir_path/test/webtest"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"
 export "FLUTTER_BUILD_NAME=1.0.0"